VMO-6541 cleanup: move ConfigFieldType to vuexBlockAndFlowHelpers (to get it exported) and use in the embedding app
