+++ src/components/interaction-designer/block-editors/GenericContactPropertyEditor.vue
@@ -123,7 +123,8 @@ import {Getter, namespace} from 'vuex-class'
 import Lang from '@/lib/filters/lang'
 import {find, has, isEmpty, map} from 'lodash'
 import {mixins} from 'vue-class-component'
-import {ConfigFieldType, isBlockInteractive} from '@/store/flow/block'
+import {isBlockInteractive} from '@/store/flow/block'
+import {ConfigFieldType} from '@/store/flow/utils/vuexBlockAndFlowHelpers'
 import VueMultiselect from 'vue-multiselect'
 import {IContactPropertyOption, IContactPropertyOptionForUISelector} from '@/store/flow/block-types/Core_SetContactPropertyStore.model'
